% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{annotateEnsembleGeneIDsWithNCBI}
\alias{annotateEnsembleGeneIDsWithNCBI}
\title{Annotate EnsEMBL GeneID with data from NCBI}
\usage{
annotateEnsembleGeneIDsWithNCBI(ids)
}
\arguments{
\item{ids}{Character strings, Ensembl GeneIDs in form of 
\code{ENS(species)(object type)(identifier).(version)}. The version is
optional.}
}
\value{
A \code{data.frame} containing following columns:
\itemize{
  \item{EnsemblID}: The input EnsemblID
  \item{GeneID}: NCBI GeneID
  \item{GeneSymbol}: Official gene symbol
  \item{GeneName}: Gene name
  \item{TaxID}: Taxonomy ID
  \item{Type}: Gene type
}
This function uses data from NCBI to annotate EnsEMBL GeneIDs. For most 
users, it is recommended to use \code{\link{annotateEnsembleGeneIDs}}, 
because it uses both data from EnsEMBL and data from NCBI to perform the
task.
}
\description{
Annotate EnsEMBL GeneID with data from NCBI
}
\details{
The \code{ncbi_gene2ensembl} collection is used.
}
\examples{
ensIDs <- readLines("inst/extdata/ribios_annotate_testdata/ensemble_geneids.txt")
ncbiAnno <- annotateEnsembleGeneIDsWithNCBI(ensIDs)
}
\seealso{
Function \code{\link{annotateEnsembleGeneIDsWithEnsembl}} annotates
EnsEMBL GeneIDs with data from Ensembl, and 
\code{\link{annotateEnsembleGeneIDs}} annotates EnsEMBL GeneIDs with both 
data from EnsEMBL and data from NCBI.
}
