% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{annotateEnsemblGeneIDsWithHumanOrtholog}
\alias{annotateEnsemblGeneIDsWithHumanOrtholog}
\title{Annotate Ensembl GeneIDs while appending human orthologs}
\usage{
annotateEnsemblGeneIDsWithHumanOrtholog(ids, multiOrth = FALSE, backend = NULL)
}
\arguments{
\item{ids}{A vector of character strings, Ensembl GeneIDs in form of 
\code{ENS(species)(object type)(identifier).(version)}. The version is
optional.}

\item{multiOrth}{Logical, whether mutliple orthologues should be returned if
exist. Deafult: \code{FALSE}}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A \code{data.frame} containing following columns:
\itemize{
  \item{EnsemblID}: The input EnsemblID
  \item{GeneID}: NCBI GeneID
  \item{GeneSymbol}: Official gene symbol
  \item{Description}: Gene description
  \item{TaxID}: Taxonomy ID
  \item{Type}: Gene type
  \item{HumanGeneID}: NCBI GeneID of the human orthologue
  \item{HumanGeneSymbol}: Official gene symbol of the human orthologue
  \item{HumanDescription}: Gene description of the human orthologue
  \item{HumanType}: Gene type of the human orthologue
}
}
\description{
Annotate Ensembl GeneIDs while appending human orthologs
}
\note{
Currently the human orthologs are looked up in NCBI. It remains to
be changed to EnsEMBL
}
\examples{
\dontrun{
ensIDs <- readLines(system.file(file.path("extdata/ribios_annotate_testdata",
                    "ensemble_geneids.txt"), package="ribiosAnnotation"))
enAnnoHumanOrt <- annotateEnsemblGeneIDsWithHumanOrtholog(ensIDs)
}
}
