% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{annotateGeneIDsWithEnsembl}
\alias{annotateGeneIDsWithEnsembl}
\title{Annotate EntrezGeneID with Ensemble identifiers}
\usage{
annotateGeneIDsWithEnsembl(
  ids,
  orthologue = FALSE,
  multiOrth = FALSE,
  type = c("gene", "transcript", "protein")
)
}
\arguments{
\item{ids}{Character or integer vector, Entrez Gene IDs. It can contain
\code{NA} or \code{NULL}.}

\item{orthologue}{Logical, whether human orthologues should be returnd}

\item{multiOrth}{Logical, whether multiple mapped orthologues should be
returned or not. Only useful when \code{orthologue} is \code{TRUE}, By
default \code{FALSE}. Be cautious with the \code{TRUE} option: in this case
the returning \code{data.frame} may have different row numbers as the input
vector.}

\item{type}{Character string, one of the following options: \code{gene}, 
\code{transcript}, and \code{protein}.}
}
\value{
A \code{data.frame} object containing the annotations.

If \code{orthologue} is set to \code{FALSE}, the data frame contains
following columns: \code{GeneID, ENSEMBL_ID, GeneSymbol, GeneName and TaxID}

If \code{orthologue} is \code{TRUE}, the data frame contains
\code{GeneID, ENSEMBL_ID, GeneSymbol,TaxID, OrigTaxID, OrigGeneID,
OrigGeneSymbol, OrigGeneName}. Note that \code{GeneID, GeneSymbol, TaxID}
contains the information of mapped orthologues, while \code{OrigXXX}
contains the information of queried genes.

If \code{multiOrth} is \code{TRUE} (only valid when \code{orthologue} is
\code{TRUE}), multiples orthologues mapped to the same gene are all
returned. This means that the result data frame may contain more rows than
the length of input vector. If set to \code{FALSE}, the resulting data frame
contains exact number of rows as the input vector length.

If either \code{transcript} or \code{protein} type is used, more than one 
items will be returned for each input ID.
}
\description{
Annotate EntrezGeneID with Ensemble identifiers
}
\examples{

options(error=utils::recover)

## normal use
annotateGeneIDsWithEnsembl(c(1234, 1432, NULL))
annotateGeneIDsWithEnsembl(c(1234, 1432, NULL), type="transcript")
annotateGeneIDsWithEnsembl(c(1234, 1432, NULL), type="protein")

annotateGeneIDsWithEnsembl(ids="22227", orthologue=TRUE)
## it is possible to mix ids of different species
annotateGeneIDsWithEnsembl(ids=c("1234", "22227"), orthologue=TRUE)

options(error=NULL)

}
\seealso{
See \code{\link{annotateEnsembl}} to annotate Ensembl IDs.

See \code{\link{annotateGeneIDs}} to get annotation for Entrez GeneIDs.

See \code{\link{annotateProbesets}} to get annotation for probesets.
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
