% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProbesets.R
\name{annotateGeneIDsWithMongoDB}
\alias{annotateGeneIDsWithMongoDB}
\alias{annotateGeneIDs}
\alias{annotateGeneSymbols}
\alias{annotateRefSeqs}
\alias{annotatemRNAs}
\title{Annotate GeneID with MongoDB}
\usage{
annotateGeneIDsWithMongoDB(ids, orthologue = FALSE, multiOrth = FALSE)
}
\arguments{
\item{ids}{Character vector, GeneIDs, GeneSymbols or mRNAs of query. It can
contain \code{NA} or \code{NULL}}

\item{orthologue}{Logical, whether human orthologues should be returned, 
currently not supported}

\item{multiOrth}{Logical, whether multiple mapped orthologues should be
returned or not. Only useful when \code{orthologue} is \code{TRUE}, By
default \code{FALSE}. Be cautious with the \code{TRUE} option: in this case
the returning \code{data.frame} may have different row numbers as the input
vector.}
}
\value{
A \code{data.frame} object containing the annotations.

If \code{orthologue} is set to \code{FALSE}, the data frame contains
following columns: \code{GeneID, GeneSymbol, GeneName and TaxID}

If \code{orthologue} is \code{TRUE}, the data frame contains \code{GeneID,
GeneSymbol,TaxID, OrigTaxID, OrigGeneID, OrigGeneSymbol, OrigGeneName}. Note
that \code{GeneID, GeneSymbol, TaxID} contains the information of mapped
orthologues, while \code{OrigXXX} contains the information of queried genes.

If \code{multiOrth} is \code{TRUE} (only valid when \code{orthologue} is
\code{TRUE}), multiples orthologues mapped to the same gene are all
returned. This means that the result data frame may contain more rows than
the length of input vector. If set to \code{FALSE}, the resulting data frame
contains exact number of rows as the input vector length.
}
\description{
Annotate GeneIDs, GeneSymbols or mRNA accesion numbers with MongoDB. This 
function will replaec annotateGeneIDs soon.
}
\note{
\code{annotatemRNAs} is an alias of \code{annotateRefSeqs}
}
\examples{

options(error=utils::recover)

## normal use
annotateGeneIDsWithMongoDB(ids=c(780, 5982, 3310))
## annotateGeneIDsWithMongoDB(ids=c(780, 5982, 3310, NA))
## annotateGeneIDsWithMongoDB(ids=c(780, 5982, 3310, NULL))
options(error=NULL)

}
\seealso{
See \code{\link{gtiChipAnnotation}} to get annotation for all
probesets in a chip.

See \code{\link{annotateProbesets}} to get annotation for probesets.
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
