% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGeneSymbols.R
\name{annotateGeneSymbolsWithHumanOrtholog}
\alias{annotateGeneSymbolsWithHumanOrtholog}
\title{Annotate GeneSymbol with human ortholog}
\usage{
annotateGeneSymbolsWithHumanOrtholog(
  ids,
  taxId,
  multiOrth = FALSE,
  backend = NULL
)
}
\arguments{
\item{ids}{Character strings, gene symbols}

\item{taxId}{Integer, NCBI taxonomy ID. Default value: 9606 (human). See \code{commonSpecies} for tax id of common species.}

\item{multiOrth}{Logical, only valid when orthologue is set to TRUE,  whether
multiple orthologues are returned}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A data.frame containing following columns:
\describe{
  \item{GeneID}{Entrez Gene ID}
  \item{GeneSymbol}{Official gene symbols}
  \item{Description}{Description}
  \item{TaxID}{NCBI Taxonomy ID}
  \item{Type}{Gene type}
  \item{HumanGeneID}{Human orthologue Entrez GeneID}
  \item{HumanGeneSymbol}{Human orthologue official gene symbol}
  \item{HumanDescription}{Human orthologue gene description}
  \item{HumanType}{Human orthologue gene type}
}
}
\description{
Annotate GeneSymbol with human ortholog
}
\examples{
\dontrun{
    annotateGeneSymbolsWithHumanOrtholog(c("Akt1", "Erbb2",
                                         "NoSuchAGene", "Tgfbr1"), 
                                         taxId=10090, multiOrth=FALSE)
}
}
