% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateHumanOrthologsWithNCBI.R
\name{annotateNonHumanGenesHumanOrthologsWithNCBI}
\alias{annotateNonHumanGenesHumanOrthologsWithNCBI}
\title{Annotate human orthologs with data from NCBI}
\usage{
annotateNonHumanGenesHumanOrthologsWithNCBI(
  geneids,
  multiOrth = FALSE,
  backend = NULL
)
}
\arguments{
\item{geneids}{Integer GeneIDs, can contain human GeneIDs.}

\item{multiOrth}{Logical, whether one gene is allowed to map to multiple
human orthologs? Default value is \code{FALSE}, i.e. only the first 
human ortholog (random choice) is returned.}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A \code{data.frame} containing following columns:
* GeneID: Input GeneID
* TaxID: Taxonomy ID of the input gene
* HumanGeneID: Human Entrez GeneID

This function annotates human orthologs for any non-human genes. Use 
\code{\link{annotateHumanOrthologsWithNCBI}} if you are not sure
whether all input GeneIDs are non-human.
}
\description{
Annotate human orthologs with data from NCBI
}
\details{
\code{ncbi_gene_info} and \code{ncbi_gene_orthologs} collections 
are used.
}
\examples{
\dontrun{
annotateNonHumanGenesHumanOrthologsWithNCBI(c(25120, 114483548, 57300, 20))
annotateNonHumanGenesHumanOrthologsWithNCBI(c(25120, 114483548, 57300, 
                                              20, 1506, 1545, 102129055))
}
}
\seealso{
\code{\link{annotateHumanOrthologsWithNCBI}}
}
