% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinGroups.R
\name{annotateProteinGroups}
\alias{annotateProteinGroups}
\title{Annotate protein groups for proteomics studies}
\usage{
annotateProteinGroups(
  ids,
  delimiter = ";",
  orthologue = FALSE,
  multiOrth = FALSE
)
}
\arguments{
\item{ids}{Character, Protein groups with identifiers 
(e.g. UniProt/SwissProt IDs) by the delimiter}

\item{delimiter}{Character, delimiter, default semicolon}

\item{orthologue}{Logical, whether human orthologues should be queried.}

\item{multiOrth}{Logical, in case of multiple orthologues, whether all of 
them should be returned
The function queries proteins in protein groups, and annotate
all proteins that cannot be annotated. For protein groups in which
no protein can be annotated, all proteins will be returned as they are,
without annotation}
}
\value{
A \code{data.frame} with following columns:
* ProteinGroup
* Protein
* GeneID
* GeneSymbol
* GeneName
* TaxID
In case \code{orthologue} is \code{TRUE}, human orthologue information is 
returned as well.
}
\description{
Annotate protein groups for proteomics studies
}
\examples{
options(error=utils::recover)
\dontrun{
annotateProteinGroups(c("A0A024RBG1;Q9NZJ9", "A0A0B4J2D5;P0DPI2",
                           "A0A0B4J2F0;A0A0U1RRL7"))
}
options(error=NULL)
}
