% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biaUser}
\alias{biaUser}
\alias{biaPwd}
\alias{bia2User}
\alias{bia2Pwd}
\alias{biaroUser}
\alias{biaroPwd}
\alias{binUser}
\alias{binPwd}
\title{Quick access of credentials}
\usage{
biaUser()

biaPwd()

bia2User()

bia2Pwd()

biaroUser()

biaroPwd()

binUser()

binPwd()
}
\description{
Quick access of credentials
}
\section{Functions}{
\itemize{
\item \code{biaPwd}: BIA password

\item \code{bia2User}: BIA2 username

\item \code{bia2Pwd}: BIA2 password

\item \code{biaroUser}: BIA READONLY username

\item \code{biaroPwd}: BIA READONLY password

\item \code{binUser}: BIN username

\item \code{binPwd}: BIN password
}}

