% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkSingleIntegerTaxId}
\alias{checkSingleIntegerTaxId}
\title{check single integer Tax ID}
\usage{
checkSingleIntegerTaxId(taxId)
}
\arguments{
\item{taxId}{Integer tax identifier, or character that can be converted to
an integer}
}
\value{
An integer tax ID if successful, otherwise the function stops and
prints error
}
\description{
check single integer Tax ID
}
