% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrLen.R
\name{chrLen}
\alias{chrLen}
\title{Get chromosome lengths in basepairs}
\usage{
chrLen(organism = c("human", "mouse", "rat"))
}
\arguments{
\item{organism}{Organism to query, supporting human, mouse and rat}
}
\value{
A \code{data.frame} containing chromosome length information.
\item{Chromosome}{Chromosome names (without the \code{CHR} prefix as in the
database} \item{Description}{Description} \item{Length}{Length of the
chromosome, in basepairs (bp)} \item{Rank}{Rank of chromosomes by lengths}
}
\description{
Get chromosome lengths in basepairs
}
\examples{

\dontrun{
chrLen("human")
chrLen("mouse")
}

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>, Isabelle Wells
<isabelle.wells@roche.com>
}
