% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxID.R
\docType{data}
\name{commonSpecies}
\alias{commonSpecies}
\title{Common species taxonomy IDs}
\format{
A \code{data.frame} containing three columns: 
\describe{
  \item{TaxID}{NCBI taxonomy ID}
  \item{ScientificName}{Scientific name}
  \item{CommonName}{Common name}
}
}
\usage{
commonSpecies
}
\description{
Common species taxonomy IDs
}
\keyword{datasets}
