% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connectMongoDB}
\alias{connectMongoDB}
\title{Connect to a MongoDB instance}
\usage{
connectMongoDB(instance = "bioinfo_read", collection = "ncbi_gene_info")
}
\arguments{
\item{instance}{Character string, the MongoDB instance to connect to}

\item{collection}{Character string, the collection to be used}
}
\value{
A pointer to a collection on the server, as returned by
 \code{\link[mongolite]{mongo}}.
}
\description{
Connect to a MongoDB instance
}
\examples{
giCon <- connectMongoDB(instance="bioinfo_read",
                        collection="ncbi_gene_info")
}
\seealso{
\code{\link{loadMongodbSecrets}}
}
