% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querydb.R
\name{formatIn}
\alias{formatIn}
\title{Formatting a vector for SQL SELECT query with IN syntax}
\usage{
formatIn(x)
}
\arguments{
\item{x}{A vector to be queried with the IN syntax}
}
\value{
A character string to be used after IN. See examples.
}
\description{
Prepare a vector for SQL SELECT query with the IN syntax
}
\examples{

myvec <- c("HH", "HM", "TH")
formatIn(myvec)
mysel <- "SELECT * FROM table WHERE city IN"
paste(mysel,formatIn(myvec))


}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
