% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxID.R
\name{getAllTaxIDs}
\alias{getAllTaxIDs}
\title{Get all taxonomy ID and scientific names offered by NCBI}
\usage{
getAllTaxIDs()
}
\value{
A data.frame containing two columns, \code{TaxID} and \code{ScientificName}.
}
\description{
Get all taxonomy ID and scientific names offered by NCBI
}
\examples{
\dontrun{
    all_tax_ids <- getAllTaxIDs()
}
}
