% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMongodbSecrets}
\alias{getMongodbSecrets}
\title{Get secrets for MongoDB connections}
\usage{
getMongodbSecrets(file = ribiosAnnotationSecretFile, instance = "bioinfo_read")
}
\arguments{
\item{file}{The secret JSON file.}

\item{instance}{String, which must be found under the \code{mongodb} section
of the JSON file}
}
\value{
A list of the following items:
\itemize{
  \item{\code{hostname}}{Hostname of the MongoDB}
  \item{\code{port}}{Port of the MongoDB}
  \item{\code{dbname}}{Database of the MongoDB}
  \item{\code{username}}{User name}
  \item{\code{password}}{Password}
}
}
\description{
Get secrets for MongoDB connections
}
\examples{
getMongodbSecrets(instance="bioinfo_read")
\dontrun{
    getMongodbSecrets(instance="decoy")
}
}
