% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gti2bioc.R
\docType{data}
\name{gtibioc}
\alias{gtibioc}
\title{Translation table between GTI and Bioconductor chip type names}
\format{
A data frame with columns:
\describe{
  \item{GTI}{GTI chip type name}
  \item{Bioconductor}{Bioconductor chip type name}
}
}
\source{
Compiled manually in December 2011.
}
\usage{
gtibioc
}
\description{
A data frame mapping GTI array names to Bioconductor array names.
}
\keyword{datasets}
