% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{guessFeatureType}
\alias{guessFeatureType}
\title{Guess feature ID type by majority voting}
\usage{
guessFeatureType(featureIDs, majority = 0.5)
}
\arguments{
\item{featureIDs}{A vector of character strings. Other input types will be 
converted to character strings.}

\item{majority}{Numeric value between 0 and 1. If the proportion of valid
feature IDs in the input matching the pattern of a certain feature type 
exceeds this value, the function returns a character string representing
the feature ID type.}
}
\value{
A character string, one of the following values: 
  \itemize{
    \item \code{GeneID}
    \item \code{GeneSymbol}
    \item \code{RefSeq}
    \item \code{Ensembl}
    \item \code{UniProt}
    \item \code{Unknown}
  }. The majority voting is done in the same order
}
\description{
Guess feature ID type by majority voting
}
\examples{
guessFeatureType(c("AKT1", "AKT2", "MAPK14"))
guessFeatureType(c(1,2,14,149))
guessFeatureType(c("NM_000259", "NM_000331"))
guessFeatureType(c("ENST00000613858.4", "ENST00000553916.5",
    "ENST00000399229.6"))
guessFeatureType(c("A2BC19", "P12345", "A0A023GPI8"))
guessFeatureType(c("CM000677.2"))
}
