% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasOracle}
\alias{hasOracle}
\title{Testing whether the Oracle client library is installed}
\usage{
hasOracle()
}
\value{
Logical. \code{TRUE} indicates Oracle SQL client is installed.
}
\description{
The function tests whether the Oracle client library is installed. If so,
the package uses \code{ROracle} as driver interface to Oracle database,
otherwise \code{RJDBC} is used.
}
\examples{

hasOracle()

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
