% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanOrthologsByTaxID.R
\name{humanOrthologsByTaxID}
\alias{humanOrthologsByTaxID}
\title{Retrieve human orthologs of genes of another species with its Taxonomy ID}
\usage{
humanOrthologsByTaxID(taxid)
}
\arguments{
\item{taxid}{An integer, a NCBI taxonomy ID to identify a species, 
for instance \code{10116} for rat, \code{10090} for mouse, and \code{9541}
for cyno (crab-eating macaque).}
}
\value{
A \code{data.frame} contains following columns:
\describe{
  \item{\code{GeneID}}{NCBI Gene ID of the query species}
  \item{\code{GeneSymbol}}{NCBI Gene symbol of the query species}
  \item{\code{Description}}{Gene description of the query species}
  \item{\code{HumanGeneID}}{NCBI Gene ID of the human homolog}
  \item{\code{HumanGeneSymbol}}{NCBI Gene symbol of the human homolog}
  \item{\code{HumanDescription}}{Gene description of the human homolog}
}
}
\description{
Retrieve human orthologs of genes of another species with its Taxonomy ID
}
\note{
To query NCBI taxonomy IDs from free-text search, visit [NCBI Taxonomy 
  Browser](https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi).
}
\examples{
\dontrun{
## human orthologs of rat genes
ratOrths <- humanOrthologsByTaxID(10116)
## human orthologs of mouse genes
mouseOrths <- humanOrthologsByTaxID(10090)
## human orthologs of cyno genes (crab-eating macaque, Macaca fascicularis)
cynoOrths <- humanOrthologsByTaxID(9541)
}
}
