% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{isValidFeatureID}
\alias{isValidFeatureID}
\title{Whether input character strings are valid feature IDs}
\usage{
isValidFeatureID(featureIDs)
}
\arguments{
\item{featureIDs}{A vector of character strings}
}
\value{
Logical vector of the same lenght as the input

Invalid feature IDs include \code{NA}, \code{"-"}, and empty string. 
Other features are deedm as valid.
}
\description{
Whether input character strings are valid feature IDs
}
\examples{
featureIDs <- c("AMPK", "", "ACTB", "-")
isValidFeatureID(featureIDs)
}
\seealso{
\code{\link{validFeatureIDs}}
}
