% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jdbcConnectString}
\alias{jdbcConnectString}
\title{Make connect string for JDBC}
\usage{
jdbcConnectString(user, password, host, port, sid)
}
\arguments{
\item{user}{A character string, user id}

\item{password}{A character string, password}

\item{host}{A character string, host name}

\item{sid}{A character string, service id}

\item{post}{A integer, port}
}
\value{
A string that can be used by \code{\link{dbConnect}} as dbname
}
\description{
Make connect string for JDBC
}
\seealso{
\code{oracleConnectString}
}
\keyword{internal}
