% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{likeGeneID}
\alias{likeGeneID}
\alias{likeRefSeq}
\alias{likeEnsembl}
\alias{likeEnsemblGeneID}
\alias{likeUniProt}
\alias{likeGeneSymbol}
\alias{likeHumanGeneSymbol}
\title{Whether input strings look like Entrez GeneIDs}
\usage{
likeGeneID(featureIDs)

likeRefSeq(featureIDs)

likeEnsembl(featureIDs)

likeEnsemblGeneID(featureIDs)

likeUniProt(featureIDs)

likeGeneSymbol(featureIDs)

likeHumanGeneSymbol(featureIDs)
}
\arguments{
\item{featureIDs}{Character strings. Input of other types are converted to them.}
}
\value{
A logical vector of the same length as input
}
\description{
Whether input strings look like Entrez GeneIDs
}
\section{Functions}{
\itemize{
\item \code{likeRefSeq()}: tests whether input strings look like NCBI RefSeq IDs

\item \code{likeEnsembl()}: tests whether input strings look like Ensembl IDs

\item \code{likeEnsemblGeneID()}: tests whether input strings look like EnsemblGeneIDs

\item \code{likeUniProt()}: tests whether input strings look like UniProt IDs

\item \code{likeGeneSymbol()}: tests whether input strings look like gene symbols

\item \code{likeHumanGeneSymbol()}: tests whether input strings look like human 
gene symbols

}}
\examples{
feats <- c("1234", "LOX", "345", "-", "", "NKX-1", "CXorf21", "Snail",
   "A2BC19", "P12345", "A0A023GPI8",
   "NM_000259", "NM_000259.3", "ENSG00000197535", "ENST00000399231.7")
likeGeneID(feats) 
likeGeneSymbol(feats)
likeRefSeq(feats)
likeEnsembl(feats)
likeUniProt(feats)
likeHumanGeneSymbol(feats)
}
\references{
Regular expression of UniProt accesion numbers is available at \url{https://www.uniprot.org/help/accession_numbers}. We requirea a whole-string match additionally

The HGNC guideline is available at \url{https://www.genenames.org/about/guidelines/}
}
