% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadMongodbSecrets}
\alias{loadMongodbSecrets}
\title{Get secrets for MongoDB connections}
\usage{
loadMongodbSecrets(file = locateSecretsFile(), instance = "bioinfo_read")
}
\arguments{
\item{file}{The secret JSON file.}

\item{instance}{String, which must be found under the \code{mongodb} section
of the JSON file}
}
\value{
A list of the following items:
\describe{
  \item{\code{hostname}}{Hostname of the MongoDB}
  \item{\code{port}}{Port of the MongoDB}
  \item{\code{dbname}}{Database of the MongoDB}
  \item{\code{username}}{User name}
  \item{\code{password}}{Password}
}
}
\description{
Get secrets for MongoDB connections
}
\examples{
loadMongodbSecrets(instance="bioinfo_read")
\dontrun{
    loadMongodbSecrets(instance="decoy")
}
}
