% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{majorityLikeHumanGeneSymbol}
\alias{majorityLikeHumanGeneSymbol}
\title{Guess the majority members of a character string look like human gene symbols}
\usage{
majorityLikeHumanGeneSymbol(x, majority = 0.8)
}
\arguments{
\item{x}{A vector of character strings}

\item{majority}{A numeric value between 0 and 1, the threshold of majority 
voting}
}
\value{
A logical value
\code{TRUE} is only returned if at least a proportion of \code{majority} 
members look like human gene symbols
}
\description{
Guess the majority members of a character string look like human gene symbols
}
\examples{
majorityLikeHumanGeneSymbol(c("AKT1", "AKT2", "MYOA")) # TRUE
majorityLikeHumanGeneSymbol(c("Akt1", "Akt2", "Myoa")) # FALSE
majorityLikeHumanGeneSymbol(c("AKT1", "Akt2", "MYOA"), majority=0.5) # TRUE
}
