% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNArefseqByTaxID.R
\name{mRNArefseqByTaxID}
\alias{mRNArefseqByTaxID}
\title{Retrieve mRNA RefSeq records by TaxID}
\usage{
mRNArefseqByTaxID(taxid = 10090)
}
\arguments{
\item{taxid}{NCBI Taxonomy ID, supported values include human (9606), mouse (10090), and rat (10116)}
}
\value{
A \code{data.frame} containing RefSeq ids and their annotation
}
\description{
Retrieve mRNA RefSeq records by TaxID
}
\examples{
\dontrun{
## get all RefSeq records of human genes
mRNArefseqByTaxID(9606)
}
}
