% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oracleConnectString}
\alias{oracleConnectString}
\title{Make connect string for Oracle}
\usage{
oracleConnectString(host, port, sid)
}
\arguments{
\item{host}{A character string, host name}

\item{sid}{A character string, service id}

\item{post}{A integer, port}
}
\value{
A string that can be used by \code{\link{dbConnect}} as dbname
}
\description{
Make connect string for Oracle
}
\seealso{
\code{jdbcConnectString}
}
\keyword{internal}
