% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAnyIDs.R
\name{removeEnsemblVersion}
\alias{removeEnsemblVersion}
\title{Remove version suffix from Ensembl IDs}
\usage{
removeEnsemblVersion(ensemblIDs)
}
\arguments{
\item{ensemblIDs}{A vector of character strings. Other types of inputs are 
converted.}
}
\value{
A character vector of the same length as input
}
\description{
Remove version suffix from Ensembl IDs
}
\examples{
ensemblIDs <- c("ENSG00000197535", "ENST00000399231.7", "ENSP00000418960.2")
removeEnsemblVersion(ensemblIDs)
}
