% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnFieldsJson}
\alias{returnFieldsJson}
\title{Construct a JSON string to indicate returned fields from a MongoDB query}
\usage{
returnFieldsJson(fields, include_id = FALSE)
}
\arguments{
\item{fields}{A vector of character strings that should be included}

\item{include_id}{Logical, whether \code{_id} should be returned. Default 
is \code{FALSE}}
}
\value{
A JSON string that represents the fields to be returned
}
\description{
Construct a JSON string to indicate returned fields from a MongoDB query
}
\examples{
returnFieldsJson(c("name", "birthday"))
returnFieldsJson(c("name", "birthday"), include_id=TRUE)
}
