% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ribiosAnnotationSecretFile}
\alias{ribiosAnnotationSecretFile}
\title{ribiosAnnotation secret file}
\format{
An object of class \code{character} of length 1.
}
\usage{
ribiosAnnotationSecretFile
}
\description{
ribiosAnnotation secret file
}
\keyword{datasets}
