% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtiChiptypes.R
\name{scriptChiptypes}
\alias{scriptChiptypes}
\title{Rscript supported chiptypes}
\usage{
scriptChiptypes()
}
\value{
A character vector of supported chip types.
}
\description{
A wrapper function for Rscripts to prompt supported chip types for
annotation
}
\details{
The functions returns a vector of \code{gtiChiptypes()}, as well as
\code{GeneID} and \code{GeneSymbol}, as the result. These are commonly used
chiptypes in Rscripts. Depending on the user input, \code{annotateAnyID} can
be called subsequently to annotate user-input features.
}
\examples{

options(error=utils::recover)
scriptChiptypes()
options(error=NULL)

}
\seealso{
\code{\link{gtiChiptypes}}, and \code{\link{annotateAnyIDs}}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
