% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortAnnotationByQuery.R
\name{sortAnnotationByQuery}
\alias{sortAnnotationByQuery}
\title{Sort the annotation table by query IDs}
\usage{
sortAnnotationByQuery(anno, ids, id_column = "GeneID", multi = FALSE)
}
\arguments{
\item{anno}{A \code{data.frame} containing annotations}

\item{ids}{A vector of character or integer, identifiers used to query
the annotation}

\item{id_column}{Character, column of the data frame where ids can be found.}

\item{multi}{In case that an identifier appears more than once in 
\code{id_column}, should all rows be returned or only the first row? Default:
\code{FALSE}, namely only the first row is returned.}
}
\value{
A \code{data.frame} sorted by the query identifiers, with the
column \code{id_column} containing exactly the same value as \code{ids}.
If the identifiers are unique and if they do not contain NA, they are used as
the row names of the data.frame; otherwise, \code{NULL} will be used.
}
\description{
Sort the annotation table by query IDs
}
\examples{
myAnno <- data.frame(GeneID=c(4,6,5), GeneName=c("Gene4", "Gene6", "Gene5"))
inputIds <- c("6", "5", "6", "4", "NotAGeneID")
sortAnnotationByQuery(myAnno, inputIds, "GeneID")

myAnno2 <- data.frame(GeneID=c(4,6,5, 5), 
                      GeneName=c("Gene4", "Gene6", "Gene5", "Gene5.V2"))
inputIds <- c("6", "5", "6", "4", "NotAGeneID")
sortAnnotationByQuery(myAnno2, inputIds, "GeneID")
sortAnnotationByQuery(myAnno2, inputIds, "GeneID", multi=TRUE)
}
