% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotByTaxID.R
\name{uniprotByTaxID}
\alias{uniprotByTaxID}
\title{Get Uniprot annotation with NCBI Taxonomy ID}
\usage{
uniprotByTaxID(taxid, orthologue = FALSE, multiOrth = FALSE)
}
\arguments{
\item{taxid}{NCBI Taxonomy ID}

\item{orthologue}{Logical, whether human orthologues should be appended to
the annotation}

\item{multiOrth}{Logical, whether to return all orthologues or the (randomly)
selected top one if multiple exist. Only valid when \code{orthologue} is set 
as \code{TRUE}.}
}
\value{
A \code{data.frame} with UniProt accessions and gene annotations.
}
\description{
Get Uniprot annotation with NCBI Taxonomy ID
}
\examples{
\dontrun{
humanUniprot <- uniprotByTaxID(9606)
}
}
\seealso{
* \code{\link{annotateUniprotAccession}}, which annotates Uniprot accessions
* \code{\link{annotateTaxID}}, which annotates genes given TaxID.
}
