% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{validFeatureIDs}
\alias{validFeatureIDs}
\title{Return valid features in a vector}
\usage{
validFeatureIDs(featureIDs)
}
\arguments{
\item{featureIDs}{A vector of character strings}
}
\value{
A filtered vector containing only valid feature IDs.

Factor input will remain factors as output, but with invalid levels dropped.
The output class will remain the same in case of integer or character input.
}
\description{
Return valid features in a vector
}
\examples{
featureIDs <- c("AMPK", "", "ACTB", "-")
validFeatureIDs(featureIDs)
}
\seealso{
\code{\link{isValidFeatureID}}
}
