% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAnyIDs.R
\name{annotateAnyIDs}
\alias{annotateAnyIDs}
\title{Annotate any identifiers}
\usage{
annotateAnyIDs(ids, orthologue = FALSE, multiOrth = FALSE, backend = NULL)
}
\arguments{
\item{ids}{A vector of identifiers. They must be of the same type. Supported 
types include Entrez GeneID, GeneSymbol, Probesets, UniProt identifiers, 
NCBI RefSeq mRNA identifiers, and Ensembl gene identifiers (with possible 
version suffixes).}

\item{orthologue}{Logical, is orthologous mapping needed?}

\item{multiOrth}{Logical, is more than one orthologs allowed}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A \code{data.frame} containing annotation information. Following
   columns exist at least: 
\enumerate{
  \item \code{Input} Input string, it will be in the first column.
  \item \code{IDType} Input ID type
  \item \code{GeneID} (Human) Entrez GeneID
  \item \code{GeneSymbol} (Human) official gene symbol
  \item \code{GeneName} (Human) gene name
  \item \code{TaxID} NCBI taxonomy ID
}
}
\description{
This annotates any identifies that can be recognized by GTI.
}
\examples{

options(error=utils::recover)

# GeneID
annotateAnyIDs(ids=c(780, 5982, 3310, NA))

# GeneSymbol
annotateAnyIDs(ids=c("DDR1", "RFC2", "HSPA6", "HSAP6"))

# Probesets
myprobes <- c("1000_at", "1004_at", "1002_f_at", "nonsense_at")
annotateAnyIDs(myprobes)

# UniProt
annotateAnyIDs(ids=c("P38398", "Q8NDF8"))

# EnsEMBL 
ensemblIDs <- c("ENSG00000197535", "ENST00000399231.7", "ENSP00000418960.2")
annotateAnyIDs(ensemblIDs)

# TODO: ENST and ENSP do not work yet
annotateAnyIDs("ENST00000399231")

# RefSeq
annotateAnyIDs(c("NM_000235", "NM_000498"))

options(error=NULL)

}
\seealso{
\code{\link{annotateGeneIDs}}, \code{\link{annotateGeneSymbols}}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
