% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{annotateEnsemblGeneIDs}
\alias{annotateEnsemblGeneIDs}
\title{Annotate Enesembl GeneIDs}
\usage{
annotateEnsemblGeneIDs(ids, orthologue = FALSE, multiOrth = FALSE)
}
\arguments{
\item{ids}{A vector of EnsemblGeneIDs in form of 
\code{ENS(species)(object type)(identifier).(version)}. The version is
optional.}

\item{orthologue}{Logical, whether human orthologues should be returned. 
Default: \code{FALSE}}

\item{multiOrth}{Logical, whether mutliple orthologues should be returned if
exist. Deafult: \code{FALSE}}
}
\value{
A \code{data.frame} object containing the annotations:
* GeneID EntrezGeneID
* GeneSymbol Official gene symbol
* Description Gene description
* TaxID Taxonomy ID
* Type Gene type

If \code{orthologue} is \code{TRUE}, following columns are appended:
* HumanGeneID
* HumanGeneSymbol
* HumanDescription
* HumanType
}
\description{
Annotate Enesembl GeneIDs
}
\examples{
\dontrun{
  annotateEnsemblGeneIDs(ids=c("ENSG00000236453", "ENSG00000170782",
                               "ENSG00000187867"))
  annotateEnsemblGeneIDs(ids=c("ENSG00000236453", "ENSG00000170782",
                               "ENSG00000187867", NA), orthologue=TRUE)
  annotateEnsemblGeneIDs(ids=c("ENSG00000174827", "ENSMUSG00000038298",
                               "ENSG00000198483", "ENSMUSG00000038354",
                               "ENSRNOG00000054947", "ENSG00000278099"),
                        orthologue=TRUE)
}
}
\seealso{
\code{\link{annotateEnsemblGeneIDsWithoutHumanOrtholog}} and
\code{\link{annotateEnsemblGeneIDsWithHumanOrtholog}}
}
