% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{annotateEnsemblGeneIDsWithEnsembl}
\alias{annotateEnsemblGeneIDsWithEnsembl}
\title{Annotate EnsEMBL GeneID with data from EnsEMBL}
\usage{
annotateEnsemblGeneIDsWithEnsembl(ids)
}
\arguments{
\item{ids}{Character strings, Ensembl GeneIDs in form of 
\code{ENS(species)(object type)(identifier).(version)}. The version is
optional.}
}
\value{
A \code{data.frame} containing following columns:
\itemize{
  \item{EnsemblID}: The input EnsemblID
  \item{GeneID}: NCBI GeneID
  \item{GeneSymbol}: Official gene symbol
  \item{Description}: Gene description
  \item{TaxID}: Taxonomy ID
}
This function uses data from EnsEMBL to annotate EnsEMBL GeneIDs. For most 
users, it is recommended to use \code{\link{annotateEnsemblGeneIDs}}, 
because it uses both data from EnsEMBL and data from NCBI to perform the
task.
}
\description{
Annotate EnsEMBL GeneID with data from EnsEMBL
}
\details{
The \code{ensembl_genes} collection is used. Note that Ensembl 
IDs often refer to novel transcripts which do not have identifiers in other
databases like NCBI Genes. If an EnsemblID is invalid or obsolete, the fields
\code{GeneName} and \code{TaxID} will be NA.
}
\examples{
\dontrun{
ensIDs <- readLines(system.file(file.path("extdata/ribios_annotate_testdata",
                    "ensemble_geneids.txt"), package="ribiosAnnotation"))
ensAnno <- annotateEnsemblGeneIDsWithEnsembl(ensIDs)
}
}
\seealso{
Function \code{\link{annotateEnsemblGeneIDsWithNCBI}} annotates
EnsEMBL GeneIDs with data from NCBI, and \code{\link{annotateEnsemblGeneIDs}}
annotates EnsEMBL GeneIDs with both data from EnsEMBL and data from NCBI.
}
