% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGeneIDs.R
\name{annotateGeneIDs}
\alias{annotateGeneIDs}
\title{Annotate Entrez GeneIDs}
\usage{
annotateGeneIDs(ids, orthologue = FALSE, multiOrth = FALSE, backend = NULL)
}
\arguments{
\item{ids}{A vector of integers or characters, encoding NCBI Entrez GeneIDs.
It can contain \code{NA} or \code{NULL}.}

\item{orthologue}{Logical, whether human orthologues should be returned. 
Default: \code{FALSE}}

\item{multiOrth}{Logical, whether mutliple orthologues should be returned if
exist. Deafult: \code{FALSE}}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A \code{data.frame} object containing the annotations:
* GeneID EntrezGeneID
* GeneSymbol Official gene symbol
* Description Gene description
* TaxID Taxonomy ID
* Type Gene type

If \code{orthologue} is \code{TRUE}, following columns are appended:
* HumanGeneID
* HumanGeneSymbol
* HumanDescription
* HumanType
}
\description{
Annotate Entrez GeneIDs
}
\examples{
\dontrun{
  annotateGeneIDs(ids=c(780, 5982, 3310))
  annotateGeneIDs(ids=c(780, 5982, 3310, NA), orthologue=TRUE)
  annotateGeneIDs(ids=c(780, 1506, 1418,
                        114483548, 57300, 
                        20, 1506, 102129055),
                        orthologue=TRUE)
}
}
\seealso{
\code{\link{annotateGeneIDsWithoutHumanOrtholog}} and
\code{\link{annotateGeneIDsWithHumanOrtholog}}
}
