% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGeneIDs.R
\name{annotateGeneIDsWithHumanOrtholog}
\alias{annotateGeneIDsWithHumanOrtholog}
\title{Annotate Entrez GeneIDs with the query of human orthologs}
\usage{
annotateGeneIDsWithHumanOrtholog(ids, multiOrth = FALSE)
}
\arguments{
\item{multiOrth}{Logical, whether mutliple orthologues should be returned if
exist. Deafult: \code{FALSE}}

\item{orthologue}{Logical, whether human orthologues should be returned. 
Default: \code{FALSE}}
}
\value{
A \code{data.frame} object containing the annotations:
* GeneID EntrezGeneID
* GeneSymbol Official gene symbol
* Description Gene description
* TaxID Taxonomy ID
* Type Gene type
* HumanGeneID
* HumanGeneSymbol
* HumanDescription
* HumanType
}
\description{
Annotate Entrez GeneIDs with the query of human orthologs
}
\examples{

\dontrun{
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310, NA))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310, NULL))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310, "1418"))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310, "NotValidGeneID"))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 3310, 1418, 5982))
  annotateGeneIDsWithHumanOrtholog(ids=c(780, 5982, 1418, 5982, 25120,
                        114483548, 57300, 20, 1506,
                        1545, 102129055))
}

}
