% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGeneIDs.R
\name{annotateGeneIDsWithoutHumanOrtholog}
\alias{annotateGeneIDsWithoutHumanOrtholog}
\title{Annotate Entrez GeneIDs without querying human orthologs}
\usage{
annotateGeneIDsWithoutHumanOrtholog(ids)
}
\arguments{
\item{ids}{A vector of integers or characters, encoding NCBI Entrez GeneIDs.
It can contain \code{NA} or \code{NULL}.}
}
\value{
A \code{data.frame} object containing the annotations:
* GeneID EntrezGeneID
* GeneSymbol Official gene symbol
* Description Gene description
* TaxID Taxonomy ID
* Type Gene type
}
\description{
Annotate Entrez GeneIDs without querying human orthologs
}
\details{
The collection {ncbi_gene_info} is used.
}
\note{
\code{annotatemRNAs} is an alias of \code{annotateRefSeqs}
}
\examples{

\dontrun{
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310, NA))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310, NULL))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310, "1418"))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310, "NotValidGeneID"))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 3310, 1418, 5982))
  annotateGeneIDsWithoutHumanOrtholog(ids=c(780, 5982, 1418, 5982, 25120,
                        114483548, 57300, 20, 1506,
                        1545, 102129055))
}

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
