% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGeneSymbols.R
\name{annotateGeneSymbols}
\alias{annotateGeneSymbols}
\title{Annotate GeneSymbols}
\usage{
annotateGeneSymbols(
  ids,
  orthologue = FALSE,
  multiOrth = FALSE,
  organism = c("human", "mouse", "rat", "any")
)
}
\arguments{
\item{ids}{Character strings, gene symbols}

\item{orthologue}{Logical, whether orthologues are to be returned}

\item{multiOrth}{Logical, only valid when orthologue is set to TRUE, whether
multiple orthologues are returned}

\item{organism}{Character string, organism
Now the function is not functional}
}
\description{
Annotate GeneSymbols
}
