% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTaxID.R
\name{annotateTaxID}
\alias{annotateTaxID}
\title{Annotations of all genes associated with the given TaxID}
\usage{
annotateTaxID(taxid)
}
\arguments{
\item{taxid}{Integer, the TaxID of the species in interest. For
instance \sQuote{9606} for Homo sapiens.}
}
\value{
A \code{data.frame} object with very similar structure as the
\code{EG_GENE_INFO} table in the database.

Rownames of the \code{data.frame} are set to \code{NULL}.
}
\description{
The function returns annotations (see details below) of all features
(probably probesets) associated with the given taxon.
}
\details{
The function reads from the backend, the MongoDB bioinfo database.
}
\examples{

\dontrun{
hsAnno <- annotateTaxID("9606")
dim(hsAnno)
head(hsAnno)

hsMtAnno <- annotateTaxID("10092")
dim(hsMtAnno)
head(hsMtAnno)
}

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
