% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateUniprotAccession.R
\name{annotateUniprotAccession}
\alias{annotateUniprotAccession}
\title{Annotate UniProt accessions or names}
\usage{
annotateUniprotAccession(accessions, orthologue = FALSE, multiOrth = FALSE)
}
\description{
Annotate UniProt accessions or names
}
\examples{
\dontrun{
annotateUniprotAccession(c("B4E0K5"))
}
}
