% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendHumanOrthologsWithNCBI.R
\name{appendHumanOrthologsWithNCBI}
\alias{appendHumanOrthologsWithNCBI}
\title{Append human orthologs to an existing annotation dataframe}
\usage{
appendHumanOrthologsWithNCBI(anno, multiOrth = FALSE)
}
\arguments{
\item{anno}{A \code{data.frame} containing at least two columns \code{GeneID}
and \code{TaxID}. The column \code{GeneID} stores Entrez GeneIDs, which are
integers (NA values and characters are tolerated). The column \code{TaxID}
stores taxonomy ID, which are integers, too (again, NA values and characters
are tolerated).}

\item{multiOrth}{Logical, whether one row is allowed to map to multiple 
orthologues}
}
\description{
Append human orthologs to an existing annotation dataframe
}
\examples{
\dontrun{
anno <- data.frame(GeneID=c(780, 1506, 114483548, 102129055, NA),
                   TaxID=c(9606, 9606, 10116, 9541, NA))
appendHumanOrthologsWithNCBI(anno)

tol_anno <- data.frame(GeneID=c(780, 1506, 114483548, 102129055, NA, "NotV"),
                   TaxID=c(9606, 9606, 10116, 9541, NA, NA))
appendHumanOrthologsWithNCBI(tol_anno)
}
}
