% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{getAnnotationBackend}
\alias{getAnnotationBackend}
\title{Get Effective Annotation Backend}
\usage{
getAnnotationBackend()
}
\value{
Character string, one of \code{"mongodb"} or
  \code{"bioconductor"}.
}
\description{
Returns the currently effective backend after resolving function argument,
option, and environment variable fallbacks.
}
