% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureID.R
\name{guessAndAnnotate}
\alias{guessAndAnnotate}
\title{Guess feature ID type by majority voting and annotate them}
\usage{
guessAndAnnotate(
  featureIDs,
  majority = 0.5,
  orthologue = FALSE,
  multiOrth = FALSE,
  taxId = 9606,
  backend = NULL
)
}
\arguments{
\item{featureIDs}{A vector of character strings. Other input types will be 
converted to character strings.}

\item{majority}{Numeric value between 0 and 1. If the proportion of valid
feature IDs in the input matching the pattern of a certain feature type 
exceeds this value, the function returns a character string representing
the feature ID type.}

\item{orthologue}{Logical, whether orthologue should be returned if the 
input features are not of human}

\item{multiOrth}{Logical, in case multiple human orthologues are available, 
should they all be returned?}

\item{taxId}{Integer, in case the input identifiers are gene 
symbols, the user can specify the organism to be used with the NCBI taxonomy ID. 
The option is passed to \code{\link{annotateGeneSymbols}}. 
Check out \code{commonSpecies} for matches between taxonomy ID and species
names.}

\item{backend}{Character string, data backend to use. One of
\code{"mongodb"} (default) or \code{"bioconductor"}.}
}
\value{
A \code{data.frame}, containing annotations of following ID types
  \itemize{
    \item \code{GeneID}
    \item \code{GeneSymbol}
    \item \code{RefSeq}
    \item \code{EnsemblGeneID}
    \item \code{Ensembl}
    \item \code{UniProt}
    \item \code{Unknown}
  }. 
In case of \code{Unknown}, a \code{data.frame} with one column (\code{FeatureName}), containing input ids, is returned.

The difference between \code{guessAndAnnotate} and \code{annotateAnyIDs} is that the later does not assume that all IDs are of the same type.
}
\description{
Guess feature ID type by majority voting and annotate them
}
\examples{
\dontrun{
  guessAndAnnotate(c("AKT1", "AKT2", "MAPK14"))
  guessAndAnnotate(c(1,2,14,149))
  guessAndAnnotate(c("NM_000259", "NM_000331"))
  guessAndAnnotate(c("ENST00000613858.4", "ENST00000553916.5",
       "ENST00000399229.6"))
  guessAndAnnotate(c("O60583", "P05997", "Q7Z624"))
  guessAndAnnotate(c("CM000677.2", "AB003434.2"))
}
}
\seealso{
\code{\link{annotateAnyIDs}}
}
