% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ribiosAnnotationSecretEnvVar}
\alias{ribiosAnnotationSecretEnvVar}
\alias{loadSecrets}
\title{Load secrets from the secret file file and set them in options}
\format{
An object of class \code{character} of length 1.
}
\usage{
ribiosAnnotationSecretEnvVar

loadSecrets(path)
}
\arguments{
\item{path}{Path to the secret file. See \code{\link{locateSecretsFile}} to
see how the program looks for the secret files if the parameter is not set.}
}
\value{
The current options of \code{ribiosAnnotation}
The function writes the \code{credentials} field of the options
After running this function, database names and passwords can be accessed.
\seealso \code{\link{locateSecretsFile}}
}
\description{
Load secrets from the secret file file and set them in options
}
\keyword{datasets}
