% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSecrets}
\alias{loadSecrets}
\title{Load secrets from a file and set them in options}
\usage{
loadSecrets(
  path = file.path("/pstore/apps/bioinfo", "ribios/secrets",
    "ribiosAnnotation-secrets.json")
)
}
\arguments{
\item{path}{Path to the secret file}
}
\value{
The current options of \code{ribiosAnnotation}
The function writes the \code{credentials} field of the options
After running this function, database names and passwords can be accessed
}
\description{
Load secrets from a file and set them in options
}
