% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ribiosAnnotationSecretFile}
\alias{ribiosAnnotationSecretFile}
\alias{loadSecrets}
\title{Load secrets from a file and set them in options}
\format{
An object of class \code{character} of length 1.
}
\usage{
ribiosAnnotationSecretFile
loadSecrets(path = ribiosAnnotationSecretFile)
}
\arguments{
\item{path}{Path to the secret file}
}
\value{
The current options of \code{ribiosAnnotation}
The function writes the \code{credentials} field of the options
After running this function, database names and passwords can be accessed
}
\description{
ribiosAnnotation needs to access databases to fetch annotations, the process
of which requires credentials for these databases. The package looks for a
file \sQuote{\code{~/.credentials/ribiosAnnotation-secrets.json}} which
contains the credentials. If this file is not found, no queries can be made.
}
\details{
A template of the credential file is provided in
\sQuote{\code{secrets/secrets-template.json}} file of the package.
}
\keyword{datasets}
