% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locateSecretsFile}
\alias{locateSecretsFile}
\title{Locate ribiosAnnotation secrets file in JSON}
\usage{
locateSecretsFile(path)
}
\arguments{
\item{path}{Path to the secret file. If not set, in case the environmental 
variable \code{RIBIOS_ANNOTATION_SECRETS_JSON} is set, its value is used as
the file path; if not, \sQuote{\code{~/.credentials/ribiosAnnotation-secrets.json}} is
used. In any case, if the file does not exist, a message will be printed.}
}
\value{
String, the normalized path of the file
}
\description{
ribiosAnnotation needs to access databases to fetch annotations, the process
of which requires credentials for these databases. The package looks for a
file in JSON format, either specified in environment variable 
\code{RIBIOS_ANNOTATION_SECRETS_JSON}, or in the file
\sQuote{\code{~/.credentials/ribiosAnnotation-secrets.json}}, which
contains the credentials. If this file is not found, no queries can be made.
}
\details{
The function locates the file and returns the normalized path of the file.
}
