% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosAnnotation-package.R
\docType{package}
\name{ribiosAnnotation-package}
\alias{ribiosAnnotation}
\alias{ribiosAnnotation-package}
\title{ribiosAnnotation: Annotation of Genes, RNAs, and Proteins in 'ribios'}
\description{
Retrieves annotation information of genomic features including genes, RNAs, and proteins from databases. It supports querying by gene identifiers, gene symbols, UniProt accessions, Ensembl identifiers, and RefSeq identifiers, as well as mapping orthologs across species using NCBI data.
}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID})

}
\keyword{internal}
