% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{setAnnotationBackend}
\alias{setAnnotationBackend}
\title{Set Global Annotation Backend}
\usage{
setAnnotationBackend(backend = c("mongodb", "bioconductor"))
}
\arguments{
\item{backend}{Character string, one of \sQuote{\code{mongodb}} or
\sQuote{\code{bioconductor}}.}
}
\value{
Invisibly returns the selected backend.
}
\description{
Set the default annotation backend for the current R session.
This affects all annotation function calls where the \code{backend}
argument is not explicitly provided.
}
